Seperate account from pet management; add drawer options for both
